requirements.txt
segment_anything
cv2import
matplotlib
torch
numpy